backend/requirements.txt
langchain
langgraph
langchain_groq
langchain_core
uvicorn 
pydantic
python-dotenv
langgraph-prebuilt
fastapi